[orchestrator] Renamed config to settings
